requirements.txt
asgiref==3.8.1
Django==4.2.11
setuptools==69.2.0
sqlparse==0.5.0
wheel==0.43.0
psycopg2==2.9.9
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2